tests/wpt/web-platform-tests/html/semantics/forms/the-button-element/button-events.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: Button - events</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#the-button-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<form name="fm1" style="display:none">
  <button id="btn">BUTTON</button>
  <button id="menu_btn" type="menu" menu="menu">MENU BUTTON</button>
</form>
<script>

var btn = document.getElementById("btn"),
    menu_btn = document.getElementById("menu_btn"),
    t1 = async_test("The submit event must be fired when click a button in submit status"),
    t2 = async_test("The reset event must be fired when click a button in reset status");

document.forms.fm1.onsubmit = t1.step_func(function (evt) {
  evt.preventDefault();
  assert_true(evt.isTrusted, "The isTrusted attribute of the submit event should be true.");
  assert_true(evt.bubbles, "The bubbles attribute of the submit event should be true.");
  assert_true(evt.cancelable, "The cancelable attribute of the submit event should be true.");
  assert_true(evt instanceof Event, "The submit event is an instance of Event interface.");
  t1.done();
});

document.forms.fm1.onreset = t2.step_func(function (evt) {
  assert_true(evt.isTrusted, "The isTrusted attribute of the reset event should be true.");
  assert_true(evt.bubbles, "The bubbles attribute of the reset event should be true.");
  assert_true(evt.cancelable, "The cancelable attribute of the reset event should be true.");
  assert_true(evt instanceof Event, "The reset event is an instance of Event interface.");
  t2.done();
});

t1.step(function () {
  btn.type = "submit";
  assert_equals(btn.type, "submit", "The button type should be 'submit'.");
  btn.click();
});

t2.step(function () {
  btn.type = "reset";
  assert_equals(btn.type, "reset", "The button type should be 'reset'.");
  btn.click();
});

</script>
